backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter==22.1
djoser
Pillow
django-colorfield
gunicorn==20.0.4
psycopg2-binary==2.9
asgiref==3.3.2
pytz==2020.1
sqlparse==0.3.1
reportlab==3.6.12